app/ingest/requirements.txt
fastapi
fastapi-limiter
asyncio==3.4.3
aiotools
aiohttp==3.8.5
python-dotenv==1.0.0
SQLAlchemy==2.0.22
uvicorn
psycopg2
cent
websockets